jit(fbw): fold in every adopted arm; identify the real MULTIFRAME flip blocker
Three things, all found by measuring shapes the corpus does not contain.

1. The live->snapshot fold ran only in the ContinueRunningNormally arm, but
   every adopted arm sets WALK_END_FLUSH_COMMITTED, and the portal epilogue
   then copies the snapshot's whole locals array onto the live root
   (`restore_resume_state_from`). A `DoneWithThisFrame*` or
   `ExitFrameWithExceptionRef` terminal therefore had frame 0's blackhole
   stores reverted, and for the exception terminal the traceback keeps the root
   frame reachable, so the stale copy is observable through `tb_frame.f_locals`
   long after the walk. Hoisted above the match.

2. Adds a frame-identity collapse guard: every recovered level must be a
   distinct frame and only `frames[0]` may be the walked frame. A level whose
   frame register resolved to the root would make the relink write an
   `f_backref` cycle and run an inner level against the root's virtualizable.
   No producer is known; the failure would be silent, so decline instead of
   relying on the absence.

3. Replaces the acceptance fixture added in the previous commit, whose header
   named the wrong mechanism. The blocker is not outer-locals staleness. The
   walker executes residuals concretely while an inline push never runs the
   interpreter's call sequence, so `ec.topframeref` still names the CALLER
   while an inlined callee body runs, and a `sys._getframe` that is itself the
   escaping residual reads the wrong frame at walk time. Adopting commits that
   answer where legacy escape/replay discards it:

       _gf().f_code.co_name   -> "main",     not "leaf"
       _gf(1).f_code.co_name  -> "<module>", not "main"
       _gf(1).f_locals["k"]   -> KeyError

   One wrong iteration per adopt, 5 adopts and 5 wrong in each part of
   `synth/getframe_while_escaping_read_frame_identity`. A `sys._getframe`
   executed after the escape, inside the blackhole, is correct, and an
   in-blackhole read of a caller local mutated earlier in the same iteration
   matches CPython and PyPy. So the original decline comment was right that an
   inline-push `enter` is the prerequisite and wrong only about which check it
   gated; that sentence is restored alongside the corrected one.

Also recorded: with a side-effecting inlined callee in a `while` loop that
returns from inside the loop, the default path runs the side effect ~5.2k times
too often -- the known trace-abort double-run class -- while the adopt produces
the exact count.

`PYRE_FBW_MULTIFRAME` stays default-off. Unit tests pass; `check.py` is 327/327
on dynasm and on cranelift.
